requirements.txt
filelock==3.29.0
fsspec==2026.3.0
ImageIO==2.37.3
Jinja2==3.1.6
joblib==1.5.3
lazy-loader==0.5
MarkupSafe==3.0.3
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.1
pillow==12.2.0
pixelate==1.0.0
pixelator==1.3.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
threadpoolctl==3.6.0
tifffile==2026.4.11
type_enforced==2.4.0
typing_extensions==4.15.0
wheel==0.46.3
